why-2025/forensics/iot-breach/pyproject.toml
[project]
name = "iot-breach"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.23.0",
]
